moved queries and mutations to queries.js
