+++ client/js/screens/AllChats/AllChatsContainer.js
@@ -2,40 +2,9 @@ import React, {Component} from 'react';
 import {Text} from 'react-native';
 import Chats from './AllChats';
 import FavesContext from '../../context/FavesContext';
-import {gql} from 'apollo-boost';
 import {Query} from '@apollo/react-components';
 import CircularLoader from '../../components/CircularLoader';
-
-export const GET_USER_CHATS = gql`
-  query($id: ID!) {
-    allConversations(filter: {members_some: {id: $id}}) {
-      id
-      members {
-        lastName
-        firstName
-        id
-        photo {
-          url
-        }
-      }
-      messages {
-        id
-        content
-        sentAt
-        recipient {
-          id
-          firstName
-          lastName
-        }
-        author {
-          id
-          firstName
-          lastName
-        }
-      }
-    }
-  }
-`;
+import {GET_USER_CHATS} from '../../config/apollo/queries';
 
 export default class AllChatsContainer extends Component {
   static navigationOptions = ({navigation}) => ({
